+++ internal/file_split/iter_test.go
@@ -1,4 +1,4 @@
-package split
+package file_split
 
 import (
 	"errors"
